Move simd control code into helper files
