+++ test/unit_tests/unit_test_image_function.cpp
@@ -3,6 +3,7 @@
 #include "../../src/filtering.h"
 #include "../../src/function_pool.h"
 #include "../../src/image_function.h"
+#include "../../src/image_function_helper.h"
 #include "../../src/image_function_simd.h"
 #include "../../src/thread_pool.h"
 #include "../../src/penguinv/cpu_identification.h"
@@ -14,27 +15,27 @@ namespace
     void PrepareFunction( const std::string& namespaceName )
     {
         if ( namespaceName == "function_pool" ) {
-            Image_Function_Simd::Simd_Activation::EnableSimd( true );
+            simd::EnableSimd( true );
             ThreadPoolMonoid::instance().resize( Unit_Test::randomValue<uint8_t>( 1, 8 ) );
         }
         else if ( namespaceName == "image_function_avx" ) {
-            Image_Function_Simd::Simd_Activation::EnableSimd( false );
-            Image_Function_Simd::Simd_Activation::EnableAvx( true );
+            simd::EnableSimd( false );
+            simd::EnableAvx( true );
         }
         else if ( namespaceName == "image_function_sse" ) {
-            Image_Function_Simd::Simd_Activation::EnableSimd( false );
-            Image_Function_Simd::Simd_Activation::EnableSse( true );
+            simd::EnableSimd( false );
+            simd::EnableSse( true );
         }
         else if ( namespaceName == "image_function_neon" ) {
-            Image_Function_Simd::Simd_Activation::EnableSimd( false );
-            Image_Function_Simd::Simd_Activation::EnableNeon( true );
+            simd::EnableSimd( false );
+            simd::EnableNeon( true );
         }
     }
 
     void CleanupFunction(const std::string& namespaceName)
     {
         if ( (namespaceName == "image_function_avx") || (namespaceName == "image_function_sse") || (namespaceName == "image_function_neon") )
-            Image_Function_Simd::Simd_Activation::EnableSimd( true );
+            simd::EnableSimd( true );
     }
 
     class FunctionRegistrator
